dom/base/test/test_bug1259588.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test for Bug 1259588</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws(new TypeError, function() {
    new File("");
  }, "new File(\"\") should throw TypeError exception");
}, "Test new File(\"\") should throw exception");
</script>
